2023.03.10 - email 수정

+++ 04. GUI/ex01-webcam_v2.py
@@ -81,16 +81,16 @@ class Ui_MainWindow(QMainWindow, form_class):
         self.pushButton_cctv_on.clicked.connect(self.live_webcam_click_on)  # 동작
         self.pushButton_cctv_off.clicked.connect(self.live_webcam_click_off) # 종료 
 
-        '''Live Webcam 공간 부분'''
-        self.box_webcam = QtWidgets.QGroupBox(self.centralwidget)
-        self.box_webcam.setGeometry(QtCore.QRect(320, 30, 771, 491))
-        font = QtGui.QFont()
-        font.setFamily("Yu Gothic UI Semibold")
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.box_webcam.setFont(font)
-        self.box_webcam.setObjectName("box_webcam")
+        # '''Live Webcam 공간 부분'''
+        # self.box_webcam = QtWidgets.QGroupBox(self.centralwidget)
+        # self.box_webcam.setGeometry(QtCore.QRect(320, 30, 771, 491))
+        # font = QtGui.QFont()
+        # font.setFamily("Yu Gothic UI Semibold")
+        # font.setPointSize(10)
+        # font.setBold(True)
+        # font.setWeight(75)
+        # self.box_webcam.setFont(font)
+        # self.box_webcam.setObjectName("box_webcam")
 
         ''' 실제 카메라 화면 출력 부분'''
         self.frame = QCameraViewfinder(self.box_webcam)
@@ -157,6 +157,26 @@ def connect_table():
         )
     return conn
 
+
+'''Count Data'''
+def countData(id= None, name= None, email= None, table_name= None, mode= None):
+    conn = connect_table()
+    cur = conn.cursor()    
+    print("mode", mode)
+    # COUNT 문
+    if mode == " SELECT":
+        SQL_COUNT = f'''SELECT COUNT(*) FROM `{table_name}`
+        WHERE `id` = '{str(id)}' OR `name`='{str(name)}' OR `email`='{str(email)}';'''
+    elif mode == " DELETE":
+        SQL_COUNT = f'''SELECT COUNT(*) FROM `{table_name}`
+        WHERE `id` = '{str(id)}' AND `name`='{str(name)}' AND `email`='{str(email)}';'''
+
+    cur.execute(SQL_COUNT)
+    count = cur.fetchone()[0]
+    conn.commit()
+    conn.close()
+
+    return count
 
 ''' Select Data'''
 def SelectData(id= None, name= None, email= None, table_name= None):
@@ -236,8 +256,8 @@ class email_form(QDialog,QWidget,form_email):
         self.name_text.clear()
         self.email_text.clear()  
 
-    def printShtname(self): # 콤보박스 상태 변경 시 터미널에 출력하는 기능 함수.
-        print(self.info_comboBox.currentText(),'모드 입니다.')
+    # def printShtname(self): # 콤보박스 상태 변경 시 터미널에 출력하는 기능 함수.
+    #     print(self.info_comboBox.currentText(),'모드 입니다.')
 
     def add_info(self):
         set_flag = False
@@ -250,20 +270,21 @@ class email_form(QDialog,QWidget,form_email):
         ## 1. 검색
         if fn_type == ' SELECT':
             if id or name or email:
-                set_flag = True
-                id_list, name_list, email_list = SelectData(
-                    id    = id, 
-                    name  = name,
-                    email = email,
-                    table_name= TABLE_CUSTOMER
-                    )
-                if len(id_list) == 0:
+                select_count = countData(id, name, email, TABLE_CUSTOMER, fn_type)
+                if select_count > 0 :
+                    set_flag = True
+                    id_list, name_list, email_list = \
+                        SelectData(
+                            id    = id, 
+                            name  = name,
+                            email = email,
+                            table_name= TABLE_CUSTOMER
+                        )
+                    QMessageBox.information(self, "Select 결과", f"{select_count}개의 데이터를 출력합니다.")
+                else:
                     QMessageBox.critical(self, "입력 오류", "입력하신 정보의 데이터가 없습니다")
-                    print("입력하신 정보의 데이터가 없습니다")
-            # close event 처럼 하나 필요 !!!
             else:
                 QMessageBox.critical(self, "입력 오류", "정보를 입력해주세요!")
-                print("ID, NAME, EMAIL중에 하나를 입력해주세요")
         
         ## 2. 입력
         elif fn_type == ' INSERT':
@@ -282,27 +303,30 @@ class email_form(QDialog,QWidget,form_email):
 
         ## 3. 삭제
         elif fn_type == ' DELETE':
-            msg = '삭제하시겠습니까?'
+            msg = '정말 삭제하시겠습니까?'
             buttonReply = QMessageBox.question(self, '삭제', msg, QMessageBox.Yes | QMessageBox.No)
 
             # 삭제 메세지에서 YES선택 시
             if buttonReply == QMessageBox.Yes:
                 # 3개의 데이터 중 하나만이라도 존재하면
                 if id and name and email:
-                    set_flag = DeleteData(
-                        id    = id, 
-                        name  = name,
-                        email = email,
-                        table_name= TABLE_CUSTOMER)
-                ## 이 부분에 팝업창 필요 !!!
+                    delete_count = countData(id, name, email, TABLE_CUSTOMER, fn_type)
+                    # delete할 행이 있다면
+                    if delete_count > 0 :
+                        set_flag = DeleteData(
+                            id    = id, 
+                            name  = name,
+                            email = email,
+                            table_name= TABLE_CUSTOMER)
+                        QMessageBox.information(self, "Delete 결과", f"{delete_count}개의 데이터를 삭제합니다.")
+                    # delete할 행이 없다면
+                    else:
+                        QMessageBox.critical(self, "입력 오류", "입력하신 정보의 데이터가 없습니다.\n데이터를 다시 확인 후 입력해주세요")
                 else:
-                    QMessageBox.critical(self, "입력 오류", "정보를 다시 입력해주세요!")
-                    print("정보를 다시 입력해주세요!")
-                    
+                    QMessageBox.critical(self, "입력 오류", "정보를 다시 입력해주세요!")     
             else: 
                 buttonReply == QMessageBox.No
                 QMessageBox.critical(self, "취소", "취소되었습니다!")
-                print('취소')
         ## info table에 입력할 데이터
         if set_flag:
             if fn_type == ' DELETE' or fn_type == ' INSERT':
